Import DSM 6.2-25556 hi3535 linux-3.4.x.txz Base on Linux Kernal v3.4.35
tar url: https://global.synologydownload.com/download/ToolChain/Synology%20NAS%20GPL%20Source/6.2-25556/hi3535/linux-3.4.x.txz
tar md5sum: 885da069d233daf06557546216199380
from: https://archive.synology.com/download/ToolChain/Synology%20NAS%20GPL%20Source/6.2-25556

+++ kernel/time/timekeeping.c
@@ -88,11 +88,8 @@ static struct timekeeper timekeeper;
  */
 __cacheline_aligned_in_smp DEFINE_SEQLOCK(xtime_lock);
 
-
 /* flag for if timekeeping is suspended */
 int __read_mostly timekeeping_suspended;
-
-
 
 /**
  * timekeeper_setup_internals - Set up internals to use clocksource clock.
@@ -197,7 +194,6 @@ static void timekeeping_update(bool clearntp)
 	update_vsyscall(&timekeeper.xtime, &timekeeper.wall_to_monotonic,
 			 timekeeper.clock, timekeeper.mult);
 }
-
 
 /**
  * timekeeping_forward_now - update clock to the current time
@@ -410,7 +406,6 @@ int do_settimeofday(const struct timespec *tv)
 
 EXPORT_SYMBOL(do_settimeofday);
 
-
 /**
  * timekeeping_inject_offset - Adds or subtracts from the current time.
  * @tv:		pointer to the timespec variable containing the offset
@@ -531,7 +526,6 @@ void getrawmonotonic(struct timespec *ts)
 	timespec_add_ns(ts, nsecs);
 }
 EXPORT_SYMBOL(getrawmonotonic);
-
 
 /**
  * timekeeping_valid_for_hres - Check if timekeeping is suitable for hres
@@ -679,7 +673,6 @@ static void __timekeeping_inject_sleeptime(struct timespec *delta)
 	update_sleep_time(timespec_add(timekeeper.total_sleep_time, *delta));
 }
 
-
 /**
  * timekeeping_inject_sleeptime - Adds suspend interval to timeekeeping values
  * @delta: pointer to a timespec delta value
@@ -713,7 +706,6 @@ void timekeeping_inject_sleeptime(struct timespec *delta)
 	/* signal hrtimers about time change */
 	clock_was_set();
 }
-
 
 /**
  * timekeeping_resume - Resumes the generic timekeeping subsystem.
@@ -978,7 +970,6 @@ static void timekeeping_adjust(s64 offset)
 				timekeeper.ntp_error_shift;
 }
 
-
 /**
  * logarithmic_accumulation - shifted accumulation of cycles
  *
@@ -1031,7 +1022,6 @@ static cycle_t logarithmic_accumulation(cycle_t offset, int shift)
 
 	return offset;
 }
-
 
 /**
  * update_wall_time - Uses the current clocksource to increment the wall time
@@ -1107,7 +1097,6 @@ static void update_wall_time(void)
 		timekeeper.ntp_error += neg << timekeeper.ntp_error_shift;
 	}
 
-
 	/*
 	 * Store full nanoseconds into xtime after rounding it up and
 	 * add the remainder to the error difference.
@@ -1165,7 +1154,6 @@ void getboottime(struct timespec *ts)
 }
 EXPORT_SYMBOL_GPL(getboottime);
 
-
 /**
  * get_monotonic_boottime - Returns monotonic time since boot
  * @ts:		pointer to the timespec to be set
@@ -1193,7 +1181,7 @@ void get_monotonic_boottime(struct timespec *ts)
 	} while (read_seqretry(&timekeeper.lock, seq));
 
 	set_normalized_timespec(ts, ts->tv_sec + tomono.tv_sec + sleep.tv_sec,
-			ts->tv_nsec + tomono.tv_nsec + sleep.tv_nsec + nsecs);
+		(s64)ts->tv_nsec + tomono.tv_nsec + sleep.tv_nsec + nsecs);
 }
 EXPORT_SYMBOL_GPL(get_monotonic_boottime);
 
@@ -1349,7 +1337,6 @@ ktime_t ktime_get_monotonic_offset(void)
 	return timespec_to_ktime(wtom);
 }
 EXPORT_SYMBOL_GPL(ktime_get_monotonic_offset);
-
 
 /**
  * xtime_update() - advances the timekeeping infrastructure
